I18N Issues based on 1.2.2
This plugin's slug is `post-draft-preview`, but the current Text Domain is `pdp`. Change the current Text Domain so it is equal to this plugin's slug, and modify the text domain in all your source files. This change is **needed**. Please refer to [this official article](https://developer.wordpress.org/plugins/internationalization/how-to-internationalize-your-plugin/#text-domains).